CSS and Corrections
A footer was added to all pages, a form was added to the results display table and the button adjusted to show correctly. Adjusted CSS for login, results table and pages about us and colors. Page colors adjusted the code inserted the UTF, footer and closed the html tag, adjust of CSS in addition to the hyperlink in the menu to return to the results page.